refactor(desktop): move wait-for-setup toggle to Experimental settings
The gate is opt-in and new, so surface it under Experimental in both UI
variants instead of Behavior/General. The toggle is extracted into a
self-contained WaitForSetupBeforeAgentSetting component (owns its query,
mutation, and optimistic update) mounted once in ExperimentalSettings, so
graduating it to a permanent section — or reverting the feature — is a
one-line mount change plus the settings-search registry entry.
