+++ apps/desktop/src/renderer/routes/_authenticated/settings/behavior/components/BehaviorSettings/BehaviorSettings.tsx
@@ -36,10 +36,6 @@ export function BehaviorSettings({ visibleItems }: BehaviorSettingsProps) {
 		SETTING_ITEM_ID.BEHAVIOR_OPEN_LINKS_IN_APP,
 		visibleItems,
 	);
-	const showWaitForSetupBeforeAgent = isItemVisible(
-		SETTING_ITEM_ID.BEHAVIOR_WAIT_FOR_SETUP_BEFORE_AGENT,
-		visibleItems,
-	);
 
 	const utils = electronTrpc.useUtils();
 
@@ -129,31 +125,6 @@ export function BehaviorSettings({ visibleItems }: BehaviorSettingsProps) {
 		},
 	);
 
-	const {
-		data: waitForSetupBeforeAgent,
-		isLoading: isWaitForSetupBeforeAgentLoading,
-	} = electronTrpc.settings.getWaitForSetupBeforeAgent.useQuery();
-	const setWaitForSetupBeforeAgent =
-		electronTrpc.settings.setWaitForSetupBeforeAgent.useMutation({
-			onMutate: async ({ enabled }) => {
-				await utils.settings.getWaitForSetupBeforeAgent.cancel();
-				const previous = utils.settings.getWaitForSetupBeforeAgent.getData();
-				utils.settings.getWaitForSetupBeforeAgent.setData(undefined, enabled);
-				return { previous };
-			},
-			onError: (_err, _vars, context) => {
-				if (context?.previous !== undefined) {
-					utils.settings.getWaitForSetupBeforeAgent.setData(
-						undefined,
-						context.previous,
-					);
-				}
-			},
-			onSettled: () => {
-				utils.settings.getWaitForSetupBeforeAgent.invalidate();
-			},
-		});
-
 	return (
 		<div className="p-6 max-w-4xl w-full">
 			<div className="mb-8">
@@ -227,34 +198,6 @@ export function BehaviorSettings({ visibleItems }: BehaviorSettingsProps) {
 							}
 							disabled={
 								isResourceMonitorLoading || setShowResourceMonitor.isPending
-							}
-						/>
-					</div>
-				)}
-
-				{showWaitForSetupBeforeAgent && (
-					<div className="flex items-center justify-between">
-						<div className="space-y-0.5">
-							<Label
-								htmlFor="wait-for-setup-before-agent"
-								className="text-sm font-medium"
-							>
-								Wait for workspace setup before starting agents
-							</Label>
-							<p className="text-xs text-muted-foreground">
-								Run the agent in the Workspace Setup terminal once setup
-								finishes instead of starting a second terminal alongside it
-							</p>
-						</div>
-						<Switch
-							id="wait-for-setup-before-agent"
-							checked={waitForSetupBeforeAgent ?? false}
-							onCheckedChange={(enabled) =>
-								setWaitForSetupBeforeAgent.mutate({ enabled })
-							}
-							disabled={
-								isWaitForSetupBeforeAgentLoading ||
-								setWaitForSetupBeforeAgent.isPending
 							}
 						/>
 					</div>
