+++ apps/desktop/src/renderer/routes/_authenticated/settings/experimental/components/ExperimentalSettings/components/WaitForSetupBeforeAgentSetting/index.ts
@@ -1,0 +1,1 @@
+export { WaitForSetupBeforeAgentSetting } from "./WaitForSetupBeforeAgentSetting";
